refactor: break backend broker cycle
